RAMagGenPy/requirements.txt
requests~=2.26.0
beautifulsoup4~=4.9.3
reportlab~=3.5.67
Pillow~=8.3.1
nltk~=3.7
numpy~=1.21.2